Each worker thread should be a separate FxA user
In general, requests by the same user can block one another, which
shows failure at very little load compared to what we actually see in
production, and in particular without causing much in the way of
server load. Therefore, rather than create one user and share it for
all the requests, create one per worker thread. This means we don't
explore the case of a single user hammering the server, but in
practice I expect most users to make only a handful of requests at a
time so as a measurement of scalability it may not be helpful.